+++ kgql-core/src/commonMain/kotlin/com/codingfeline/kgql/core/KgqlRequestBody.kt
@@ -1,14 +1,10 @@
 package com.codingfeline.kgql.core
 
-import kotlinx.serialization.Optional
-import kotlinx.serialization.Serializable
-
 /**
- * Root class for GraphQL request body
+ * Root interface for GraphQL request body
  */
-@Serializable
-data class KgqlRequestBody<T>(
-    @Optional val operationName: String? = null,
-    val query: String,
-    @Optional val variables: T? = null
-)
+interface KgqlRequestBody<T> {
+    val operationName: String?
+    val query: String
+    val variables: T?
+}